backend/requirement.txt
fastapi
uvicorn
pydantic
requests
beautifulsoup4
playwright
python-dotenv
